Add files via upload
This is a driver from  https://github.com/oh1ko/am2301.
The original driver would not be compiled because of changes in the Linux kernel. I made some small modifications to be able to compile the driver.
You can install the driver directly if you don't want to compile it yourself.